editor webpack merge
