+++ editor/webpack.production.config.js
@@ -1,43 +1,16 @@
 const path = require('path');
-const VueLoaderPlugin = require('vue-loader/lib/plugin');
 
-const vueLoader = {
-    test: /\.(vue)$/,
-    exclude: /node_modules/,
-    use: 'vue-loader'
-};
+const baseconfig = require('./webpack.base.config');
 
-const scssLoader = {
-    test: /\.(css|scss)$/,
-    use: ['css-loader', 'sass-loader']
-};
+const merge = require('webpack-merge');
+const webpack = require('webpack');
 
-module.exports = {
-    entry: './src/index.js',
+const config = {
     mode: 'production',
     output: {
         path: path.resolve(__dirname, 'dist'),
-        filename: 'editor.release.js'
-    },
-    module: {
-        rules: [vueLoader, scssLoader]
-    },
-    externals: {
-        electron: 'window.require("electron")',
-        fs: 'window.require("fs")',
-        util: 'window.require("util")',
-        process: 'require("process")'
-    },
-    resolve: {
-        alias: {
-            vue$: path.resolve('..', 'node_modules', 'vue', 'dist', 'vue.esm.js')
-        },
-        modules: [
-            path.resolve('..', 'node_modules'),
-            path.resolve('..', 'common', 'modules')
-        ]
-    },
-    plugins: [
-        new VueLoaderPlugin()
-    ]
+        filename: 'editor.js'
+    }
 };
+
+module.exports = merge(baseconfig, config);
